backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
httpx>=0.28.0
numpy>=2.2.0
scikit-learn>=1.6.0
xgboost>=2.1.0
pandas>=2.2.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
